renames Utilities class to DataExportUtilities
